Zum Commit vorgemerkte Änderungen: 	geändert:       Movim/build.gradle 	gelöscht:       Movim/google-services.json 	geändert:       Movim/src/main/AndroidManifest.xml 	geändert:       Movim/src/main/java/com/movim/movim/MainActivity.java 	gelöscht:       Movim/src/main/java/com/movim/movim/NotificationService.java 	geändert:       Movim/src/main/java/com/movim/movim/VisioActivity.java 	geändert:       Movim/src/main/res/layout/activity_main.xml 	geändert:       Movim/src/main/res/layout/activity_visio.xml 	geändert:       build.gradle 	neue Datei:     fastlane/metadata/android/en-US/changelog/14.txt 	gelöscht:       fastlane/metadata/android/en-US/changelog/15.txt
