+++ Movim/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'com.android.application'
-apply plugin: 'com.google.gms.google-services'
 
 android {
     compileSdkVersion 30
@@ -8,8 +7,8 @@ android {
         applicationId "com.movim.movim"
         minSdkVersion 21
         targetSdkVersion 30
-        versionCode 15
-        versionName "0.19"
+        versionCode 16
+        versionName "0.19.0.0"
     }
 
     buildTypes {
@@ -22,8 +21,6 @@ android {
 
 dependencies {
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.appcompat:appcompat:1.0.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation platform('com.google.firebase:firebase-bom:26.3.0')
-    implementation 'com.google.firebase:firebase-messaging'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
 }
