--- Movim/google-services.json
@@ -1,39 +1,0 @@
-{
-  "project_info": {
-    "project_number": "514406907830",
-    "project_id": "movim-official-pod",
-    "storage_bucket": "movim-official-pod.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:514406907830:android:34e10aba8bf4b4e1768b17",
-        "android_client_info": {
-          "package_name": "com.movim.movim"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "514406907830-2mfu2nrkg6vd581voack04f00puj8asu.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyCpMgI7JeZVFzdBntUOu9WNqMjGvlPHmmI"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "514406907830-2mfu2nrkg6vd581voack04f00puj8asu.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
