+++ Movim/src/main/AndroidManifest.xml
@@ -19,32 +19,12 @@
         android:roundIcon="@mipmap/ic_logo_round"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
-        <service
-            android:name=".NotificationActivity"
-            android:enabled="true"
-            android:exported="true"></service>
-        <service
-            android:name=".NotificationService"
-            android:enabled="true"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
-
-        <meta-data
-            android:name="firebase_analytics_collection_enabled"
-            android:value="false" />
-
-        <activity
-            android:name=".ShareActivity"
+        <activity android:name=".ShareActivity"
             android:label="@string/share_new_post"
             android:launchMode="singleTask">
             <intent-filter>
                 <action android:name="android.intent.action.SEND" />
-
                 <category android:name="android.intent.category.DEFAULT" />
-
                 <data android:mimeType="text/plain" />
             </intent-filter>
         </activity>
@@ -52,21 +32,18 @@
             android:name=".VisioActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/call"
-            android:launchMode="standard"
+            android:theme="@style/SplashTheme"
             android:taskAffinity=".MainActivity"
-            android:theme="@style/SplashTheme" />
+            android:launchMode="standard" />
         <activity
             android:name=".MainActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/app_name"
-            android:launchMode="singleInstance"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
