+++ Movim/src/main/java/com/movim/movim/MainActivity.java
@@ -1,6 +1,11 @@
 package com.movim.movim;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 import android.Manifest;
@@ -8,30 +13,38 @@ import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.AlertDialog.Builder;
-import android.content.pm.PackageManager;
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+
 import androidx.annotation.NonNull;
 import androidx.core.app.ActivityCompat;
+import androidx.core.app.NotificationCompat;
+import androidx.core.content.ContextCompat;
 
+import android.app.PendingIntent;
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.IntentFilter;
+import android.content.pm.PackageManager;
 import android.content.res.Configuration;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.graphics.Color;
-import android.graphics.PixelFormat;
 import android.net.Uri;
-import android.net.http.SslError;
+//import android.net.http.SslError;
 import android.os.Build;
 import android.os.Bundle;
-import androidx.core.content.ContextCompat;
 import android.text.InputType;
-import android.util.Log;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.Window;
-import android.view.WindowManager;
 import android.webkit.HttpAuthHandler;
 import android.webkit.JavascriptInterface;
 import android.webkit.PermissionRequest;
-import android.webkit.SslErrorHandler;
+//import android.webkit.SslErrorHandler;
 import android.webkit.ValueCallback;
 import android.webkit.WebChromeClient;
 import android.webkit.WebView;
@@ -41,32 +54,27 @@ import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.messaging.FirebaseMessaging;
-
-import static com.google.firebase.messaging.Constants.TAG;
-
 @SuppressLint("SetJavaScriptEnabled")
 public class MainActivity extends Activity {
-	private WebView webview;
-	private ProgressBar progressbar;
-	private String firebaseToken = "";
-	private static MainActivity instance;
-
-	private ValueCallback<Uri[]> mUploadMessageArray;
 	private final static int FILE_REQUEST_CODE = 1;
 	private final static int CAMERA_REQUEST_CODE = 2;
+	private WebView webview;
+	private ProgressBar progressbar;
+	private HashMap<String, List<String>> notifs;
+	@SuppressLint("StaticFieldLeak")
+	private static MainActivity instance;
+
+	private ValueCallback<Uri> mUploadMessage;
+	private ValueCallback<Uri[]> mUploadMessageArray;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		setTheme(R.style.SplashTheme);
+		this.notifs = new HashMap<String, List<String>>();
 
 		super.onCreate(savedInstanceState);
-
 		getWindow().requestFeature(Window.FEATURE_PROGRESS);
 		getWindow().requestFeature(Window.FEATURE_NO_TITLE);
-		getWindow().setFormat(PixelFormat.RGBA_8888);
 		setContentView(R.layout.activity_main);
 
 		webview = findViewById(R.id.webview);
@@ -74,21 +82,8 @@ public class MainActivity extends Activity {
 		webview.getSettings().setJavaScriptEnabled(true);
 		webview.getSettings().setDomStorageEnabled(true);
 		webview.getSettings().setMixedContentMode(0);
-		webview.getSettings().setAppCacheEnabled(true);
 		webview.getSettings().setMediaPlaybackRequiresUserGesture(false);
 
-		FirebaseMessaging.getInstance().getToken()
-				.addOnCompleteListener(new OnCompleteListener<String>() {
-					@Override
-					public void onComplete(@NonNull Task<String> task) {
-						if (!task.isSuccessful()) {
-							Log.w(TAG, "Fetching FCM registration token failed", task.getException());
-							return;
-						}
-						firebaseToken = task.getResult();
-					}
-				});
-		
 		if (Build.VERSION.SDK_INT >= 21) {
 			webview.getSettings().setAllowUniversalAccessFromFileURLs(true);
 			getWindow().setNavigationBarColor(Color.parseColor("#000000"));
@@ -102,42 +97,47 @@ public class MainActivity extends Activity {
 			webview.setLayerType(View.LAYER_TYPE_SOFTWARE, null);
 		}
 
-		progressbar = findViewById(R.id.progress);
+		progressbar = (ProgressBar) findViewById(R.id.progress);
 		progressbar.setIndeterminate(true);
 
 		webview.addJavascriptInterface(this, "Android");
-		webview.setBackgroundColor(Color.TRANSPARENT);
-		webview.setVisibility(View.GONE);
 		webview.setWebChromeClient(new WebChromeClient() {
-			@Override
-			public void onPermissionRequest(PermissionRequest request) {
-				checkAndRequestPermissions();
-				request.grant(request.getResources());
-			}
-
 			public void onProgressChanged(WebView view, int progress) {
+				progressbar.setProgress(progress);
 				if (progress < 100 && progress > 0 && progressbar.getVisibility() == ProgressBar.GONE) {
 					progressbar.setIndeterminate(true);
 					progressbar.setVisibility(ProgressBar.VISIBLE);
 				}
-
-				if (progress > 80) {
-					webview.setVisibility(View.VISIBLE);
-				}
-
 				if (progress == 100) {
 					progressbar.setVisibility(ProgressBar.GONE);
-
-					if (!firebaseToken.isEmpty()) {
-						webview.loadUrl("javascript:Presence.setFirebaseToken('" + firebaseToken + "')");
-					}
 				}
+
 			}
 
-			public boolean onShowFileChooser(WebView webView, ValueCallback<Uri[]> uploadMessageArray, WebChromeClient.FileChooserParams fileChooserParams) {
-				if (mUploadMessageArray != null) {
+			public void onCloseWindow(WebView view){
+				super.onCloseWindow(view);
+			}
+
+			public void openFileChooser(ValueCallback<Uri> uploadMsg) {
+				openFileChooser(uploadMsg, null, null);
+			}
+
+			public void openFileChooser(ValueCallback uploadMsg, String acceptType) {
+				openFileChooser(uploadMsg, null, null);
+			}
+
+			public void openFileChooser(ValueCallback<Uri> uploadMsg, String acceptType, String capture) {
+				mUploadMessage = uploadMsg;
+				Intent i = new Intent(Intent.ACTION_GET_CONTENT);
+				i.addCategory(Intent.CATEGORY_OPENABLE);
+				i.setType("*/*");
+				MainActivity.this.startActivityForResult(Intent.createChooser(i, "File Browser"), FILE_REQUEST_CODE);
+			}
+
+			public boolean onShowFileChooser(WebView webView, ValueCallback<Uri[]> uploadMessageArray,
+											 WebChromeClient.FileChooserParams fileChooserParams) {
+				if (mUploadMessageArray != null)
 					mUploadMessageArray.onReceiveValue(null);
-				}
 
 				mUploadMessageArray = uploadMessageArray;
 
@@ -150,6 +150,12 @@ public class MainActivity extends Activity {
 				ci.putExtra(Intent.EXTRA_TITLE, "File Browser");
 				startActivityForResult(ci, FILE_REQUEST_CODE);
 				return true;
+			}
+
+			@Override
+			public void onPermissionRequest(final PermissionRequest request) {
+				checkAndRequestPermissions();
+				request.grant(request.getResources());
 			}
 		});
 
@@ -177,17 +183,18 @@ public class MainActivity extends Activity {
 				return true;
 			}
 
-			public void onReceivedSslError(WebView view, final SslErrorHandler handler, SslError error) {
-				webview.loadUrl("file:///android_asset/ssl.html");
-			}
+			//deactivated to avoid connection issues to some servers
+			//public void onReceivedSslError(WebView view, final SslErrorHandler handler, SslError error) {
+			//	webview.loadUrl("file:///android_asset/ssl.html");
+			//}
 
 			public void onReceivedHttpAuthRequest(final WebView view, final HttpAuthHandler handler, final String host,
-					final String realm) {
+												  final String realm) {
 				final String[] httpAuth = new String[2];
 				final String[] viewAuth = view.getHttpAuthUsernamePassword(host, realm);
 				final EditText usernameInput = new EditText(MainActivity.getInstance());
 				final EditText passwordInput = new EditText(MainActivity.getInstance());
- 
+
 				httpAuth[0] = viewAuth != null ? viewAuth[0] : new String();
 				httpAuth[1] = viewAuth != null ? viewAuth[1] : new String();
 
@@ -225,11 +232,6 @@ public class MainActivity extends Activity {
 		instance = this;
 	}
 
-	@Override
-	protected void onStop () {
-		super.onStop() ;
-		startService( new Intent( this, NotificationService. class )) ;
-	}
 
 	private boolean checkAndRequestPermissions() {
 		int permissionCamera = ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA);
@@ -266,7 +268,7 @@ public class MainActivity extends Activity {
 	@Override
 	public void onNewIntent(Intent intent) {
 		if (intent.getAction() != null) {
-			//this.notifs.remove(intent.getAction());
+			this.notifs.remove(intent.getAction());
 			webview.loadUrl(intent.getAction());
 		}
 	}
@@ -275,44 +277,32 @@ public class MainActivity extends Activity {
 	@Override
 	public void onConfigurationChanged(Configuration newConfig) {
 		super.onConfigurationChanged(newConfig);
-	};
+	}
 
 	public static MainActivity getInstance() {
 		return instance;
 	}
 
 	@Override
-	protected void onResume() {
-		super.onResume();
-		if (mUploadMessageArray == null)
-			return;
-
-		mUploadMessageArray.onReceiveValue(new Uri[]{});
-		mUploadMessageArray = null;
-	}
-
-	@Override
 	protected void onActivityResult(int requestCode, int resultCode, Intent data) {
 		super.onActivityResult(requestCode, resultCode, data);
 
-		if (requestCode != FILE_REQUEST_CODE || mUploadMessageArray == null) {
+		if (requestCode != FILE_REQUEST_CODE || resultCode != Activity.RESULT_OK || data.getData() == null)
 			return;
-		}
 
-		Uri[] results = null;
-		if (resultCode == Activity.RESULT_OK) {
-			String dataString = data.getDataString();
-			if (dataString != null) {
-				results = new Uri[]{Uri.parse(dataString)};
-			}
-		}
-		mUploadMessageArray.onReceiveValue(results);
-		mUploadMessageArray = null;
-	}
+		if (Build.VERSION.SDK_INT >= 21) {
+			if (mUploadMessageArray == null)
+				return;
 
-	@JavascriptInterface
-	public void clearFirebaseToken() {
-		firebaseToken = "";
+			mUploadMessageArray.onReceiveValue(new Uri[] { data.getData() });
+			mUploadMessageArray = null;
+		} else {
+			if (mUploadMessage == null)
+				return;
+
+			mUploadMessage.onReceiveValue(data.getData());
+			mUploadMessage = null;
+		}
 	}
 
 	@JavascriptInterface
@@ -330,24 +320,107 @@ public class MainActivity extends Activity {
 
 	@JavascriptInterface
 	public void clearNotifications(String action) {
-		Intent notificationService = new Intent(this, NotificationService.class);
-		notificationService.setAction("clear");
-		notificationService.putExtra("action", action);
-		startService(notificationService);
+		NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+		notificationManager.cancel(action, 0);
+
+		if (this.notifs.get(action) != null) {
+			this.notifs.remove(action);
+		}
+
+		this.updateNotifications();
 	}
 
 	@JavascriptInterface
 	public void showNotification(String title, String body, String picture, String action) {
-		Intent notificationService = new Intent(this, NotificationService.class);
-		notificationService.setAction("notify");
-		notificationService.putExtra("title", title);
-		notificationService.putExtra("body", body);
-		notificationService.putExtra("picture", picture);
-		notificationService.putExtra("action", action);
-		startService(notificationService);
+		Bitmap pictureBitmap = getBitmapFromURL(picture);
+
+		Intent i = new Intent(this, MainActivity.class);
+		if (action != null) {
+			i.setAction(action);
+			i.addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP | Intent.FLAG_ACTIVITY_CLEAR_TOP);
+		}
+		PendingIntent pi = PendingIntent.getActivity(this, 0, i, PendingIntent.FLAG_UPDATE_CURRENT);
+
+		// The deleteIntent declaration
+		Intent deleteIntent = new Intent(action);
+		PendingIntent pendingDeleteIntent = PendingIntent.getBroadcast(this, 0, deleteIntent,
+				PendingIntent.FLAG_UPDATE_CURRENT);
+		registerReceiver(receiver, new IntentFilter(action));
+
+		// Integer counter;
+		List<String> messages;
+
+		// There is already pending notifications
+		if (this.notifs.get(action) != null) {
+			messages = this.notifs.get(action);
+		} else {
+			messages = new ArrayList<String>();
+		}
+
+		messages.add(body);
+
+		if (messages.size() > 5) {
+			messages.remove(0);
+		}
+
+		this.notifs.put(action, messages);
+
+		// We create the inbox
+		NotificationCompat.InboxStyle style = new NotificationCompat.InboxStyle();
+		for (int j = 0; j < messages.size(); j++) {
+			style.addLine(messages.get(j));
+		}
+
+		style.setBigContentTitle(title);
+
+		NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+
+		String channelId = "channel-movim";
+		String channelName = "Movim";
+		String groupId = "movim";
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+			NotificationChannel mChannel = new NotificationChannel(channelId, channelName, NotificationManager.IMPORTANCE_DEFAULT);
+			notificationManager.createNotificationChannel(mChannel);
+		}
+
+		Notification notification = new NotificationCompat.Builder(this, channelId)
+				.setSmallIcon(R.drawable.ic_stat_name)
+				.setLargeIcon(pictureBitmap)
+				.setContentTitle(title)
+				.setContentText(body)
+				.setContentIntent(pi)
+				.setDeleteIntent(pendingDeleteIntent).setAutoCancel(true).setColor(Color.parseColor("#3F51B5"))
+				.setLights(Color.parseColor("#3F51B5"), 1000, 5000)
+				.setNumber(messages.size())
+				.setStyle(style)
+				.setGroup(groupId)
+				.build();
+		notificationManager.notify(action, 0, notification);
+
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+			Notification summaryNotification = new NotificationCompat.Builder(this, channelId)
+					.setSmallIcon(R.drawable.ic_stat_name)
+					.setLargeIcon(BitmapFactory.decodeResource(getResources(), R.drawable.ic_vectorial))
+					.setGroup(groupId)
+					.setGroupSummary(true)
+					.setAutoCancel(true)
+					.build();
+			notificationManager.notify("summary", 0, summaryNotification);
+		}
 	}
 
-	/*private final BroadcastReceiver receiver = new BroadcastReceiver() {
+	protected void updateNotifications() {
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+			NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+			int counter = notificationManager.getActiveNotifications().length;
+
+			if (counter <= 1) {
+				notificationManager.cancel("summary", 0);
+			}
+		}
+	}
+
+	private final BroadcastReceiver receiver = new BroadcastReceiver() {
 		@Override
 		public void onReceive(Context context, Intent intent) {
 			if (intent.getAction() != null) {
@@ -355,7 +428,22 @@ public class MainActivity extends Activity {
 			}
 			unregisterReceiver(this);
 		}
-	};*/
+	};
+
+	public static Bitmap getBitmapFromURL(String src) {
+		try {
+			URL url = new URL(src);
+			HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+			connection.setDoInput(true);
+			connection.connect();
+			InputStream input = connection.getInputStream();
+			Bitmap myBitmap = BitmapFactory.decodeStream(input);
+			return myBitmap;
+		} catch (IOException e) {
+			e.printStackTrace();
+			return null;
+		}
+	}
 
 	public boolean onKeyUp(int keyCode, KeyEvent event) {
 		if (keyCode == KeyEvent.KEYCODE_MENU) {
