+++ Movim/src/main/res/layout/activity_visio.xml
@@ -2,7 +2,6 @@
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
     android:orientation="vertical"
-    android:layout_marginTop="-24dp"
     android:fitsSystemWindows="true">
 
     <WebView
