+++ fastlane/metadata/android/en-US/changelog/14.txt
@@ -1,0 +1,13 @@
+* Fits with the 0.19 Movim release
+* update gradle plugin
+* update to targetSdkVersion 30
+* update compileSdkVersion 30
+* versionCode 14
+* versionName "0.19.0.0"
+* update com.android.support.constraint:constraint-layout to 2.0.4
+* deactivated ssl checkup to avoid connections issues to some movim instances
+* migrated to AndroidX
+* improved permissions for Microphone and Camera
+* little improvements of notifications 
+* corrected screen size to fit under top bar
+* little code cleanup
